# docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: SuperCore API
  version: 1.0.0
  description: |
    # SuperCore - Universal Entity Management Platform API

    SuperCore is a revolutionary meta-platform for creating Core Banking systems through natural language.

    ## Core Concepts

    - **Oracle**: Platform consciousness that knows its identity, licenses, and capabilities
    - **Object Definitions**: Blueprints (DNA) for all entities in the system
    - **Instances**: Living entities created from object definitions
    - **Relationships**: Connections between instances forming a semantic graph
    - **Validation Rules**: Reusable business rules and compliance checks

    ## Authentication

    Currently in development. Future versions will use Logto for OAuth 2.0/OIDC.

    ## Rate Limiting

    Default: 100 requests per minute per IP

    ## Versioning

    API version is included in the URL path: `/api/v1/`

  contact:
    name: SuperCore Team
    email: dev@lbpay.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api-dev.supercore.lbpay.com/api/v1
    description: Development environment
  - url: https://api.supercore.lbpay.com/api/v1
    description: Production environment
...
paths:
...
  /instances:
    get:
      tags: [Instances]
      summary: List instances
      description: Returns a paginated list of instances
      operationId: listInstances
      parameters:
        - name: object_definition_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by object definition
        - name: current_state
          in: query
          schema:
            type: string
          description: Filter by current state
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of instances
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Instance'
                  total:
                    type: integer

    post:
      tags: [Instances]
      summary: Create a new instance
      description: Creates a new instance from an object definition
      operationId: createInstance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInstance'
            examples:
              conta_corrente:
                summary: Create a checking account instance
                value:
                  object_definition_id: 550e8400-e29b-41d4-a716-446655440000
                  data:
                    cpf: "12345678901"
                    nome: "João Silva"
                    saldo: 1000.00
      responses:
        '201':
          description: Instance created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instance'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /instances/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Instances]
      summary: Get instance by ID
      operationId: getInstance
      responses:
        '200':
          description: Instance found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instance'
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags: [Instances]
      summary: Update instance data
      operationId: updateInstance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInstance'
      responses:
        '200':
          description: Instance updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instance'

    delete:
      tags: [Instances]
      summary: Delete instance
      description: Soft delete an instance
      operationId: deleteInstance
      responses:
        '204':
          description: Instance deleted
  /instances/{id}/transition:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags: [Instances]
      summary: Transition instance state
      description: Moves instance to a new state using the FSM
      operationId: transitionInstanceState
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to_state]
              properties:
                to_state:
                  type: string
                  example: ACTIVE
                reason:
                  type: string
                  example: KYC verification completed
      responses:
        '200':
          description: State transition successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instance'
        '400':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /instances/{id}/history:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Instances]
      summary: Get instance state history
      description: Returns the complete state transition history
      operationId: getInstanceHistory
      responses:
        '200':
          description: State history
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    state:
                      type: string
                    timestamp:
                      type: string
                      format: date-time
                    user:
                      type: string
                    reason:
                      type: string
...
components:
  schemas:
...
    Instance:
      type: object
      properties:
        id:
          type: string
          format: uuid
        object_definition_id:
          type: string
          format: uuid
        data:
          type: object
        current_state:
          type: string
        state_history:
          type: array
        version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_deleted:
          type: boolean
    CreateInstance:
      type: object
      required: [object_definition_id, data]
      properties:
        object_definition_id:
          type: string
          format: uuid
        data:
          type: object
    UpdateInstance:
      type: object
      required: [data]
      properties:
        data:
          type: object
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object